Abort if there are too many colors
The source PNG file must not contain more palette entries
than allowed by the selected bit depth. Exit in error
when this happens rather than outputting an oversized palette.

I suppose there *could* be valid uses cases for having a larger palette.
For instance, you could want to have a dark/night version of the first
palette. When and if this becomes necessary, a new command-line option
to allow this can easily added.